backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.4.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0